requirements.txt
anaconda==0.0.1.1
asn1crypto==0.24.0
atomicwrites==1.4.0
attrs==19.3.0
awscli==1.16.303
awsebcli==3.16.0
backcall==0.1.0
bleach==3.1.0
boto==2.49.0
boto3==1.14.12
botocore==1.17.12
cement==2.8.2
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
colorama==0.4.3
colour==0.1.5
conda==4.3.13
cryptography==2.4.2
cycler==0.10.0
Cython==0.29.14
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.8
django-bootstrap3==11.0.0
django-registration-redux==2.5
django-smtp-ssl==1.0
docutils==0.15.2
entrypoints==0.3
frozendict==1.2
future==0.16.0
graphviz==0.11
hellowebapp-deploy==1.0.2
idna==2.7
importlib-metadata==1.6.0
ipykernel==5.1.0
ipyparallel==6.2.4
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
more-itertools==8.3.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==1.11
notebook==5.7.8
numexpr==2.7.1
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
paramz==0.9.5
parso==0.4.0
pathspec==0.5.9
patsy==0.5.1
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.5
py==1.8.1
pyasn1==0.4.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.4.7
pyrsistent==0.14.11
PySocks==1.6.8
pytest==5.4.2
pytexit==0.3.3
python-dateutil==2.7.5
python-decouple==3.1
pytz==2020.1
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.20.1
rsa==3.4.2
ruamel-yaml==0.15.46
ruamel.yaml==0.15.94
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
semantic-version==2.5.0
Send2Trash==1.5.0
six==1.11.0
SQLAlchemy==1.3.3
sqlparse==0.3.1
statsmodels==0.9.0
tables==3.6.1
tdda==1.0.13
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.46.0
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
waitress==1.1.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==3.2.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==4.6.0
